Capstone_Fall2025/src/Demo/requirements.txt
streamlit>=1.47.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
Pillow>=9.5.0
xgboost>=1.7.0
scikit-learn>=1.3.0
shap>=0.42.0
textblob>=0.17.0
nltk>=3.8.0
textstat>=0.7.0
boto3>=1.26.0